+++ src/binaryapi/ActiveSymbols.js
