+++ src/binaryapi/__tests__/DummyBinaryAPI.js
