+++ src/binaryapi/__tests__/TradingTimes.spec.js
