+++ src/binaryapi/index.js
@@ -1,0 +1,9 @@
+import ActiveSymbols from './ActiveSymbols';
+import TradingTimes from './TradingTimes';
+import BinaryAPI from './BinaryAPI';
+
+export {
+    ActiveSymbols,
+    TradingTimes,
+    BinaryAPI,
+};
