+++ src/feed/index.js
@@ -1,9 +1,5 @@
 import Feed from './Feed';
-import BinaryAPI from './BinaryAPI';
-import TradingTimes from './TradingTimes';
 
 export {
     Feed,
-    BinaryAPI,
-    TradingTimes,
 };
