+++ src/store/ChartStore.js
@@ -4,8 +4,8 @@ import PendingPromise from '../utils/PendingPromise';
 import Context from '../components/ui/Context';
 import KeystrokeHub from '../components/ui/KeystrokeHub';
 import '../components/ui/Animation';
-import { BinaryAPI, Feed, TradingTimes } from '../feed';
-import ActiveSymbols from './ActiveSymbols';
+import { Feed } from '../feed';
+import { ActiveSymbols, BinaryAPI, TradingTimes } from '../binaryapi';
 import { calculateTimeUnitInterval, getUTCDate, cloneCategories } from '../utils';
 
 CIQ.ChartEngine.prototype.createYAxisLabel = function (panel, txt, y, backgroundColor, color, ctx, yAxis) {
